+++ packages/mobile/src/cells/CellDetail.tsx
@@ -8,14 +8,18 @@ export type CellDetailVariant = 'foregroundMuted' | 'negative' | 'positive' | 'w
 export type CellDetailProps = {
   /**
    * Label and/or extra detail. This prop is only intended to accept a string or Text component;
-   * other use cases, while allowed, are not supported and may result in unexpected behavior.
+   * other use cases, while allowed, are not supported and may result in unexpected behavior. For arbitrary content, use `detailNode`.
    */
   detail?: React.ReactNode;
+  /** React node to render label and/or extra detail. Takes precedence over `detail`. */
+  detailNode?: React.ReactNode;
   /**
    * Subdetail providing more information. This prop is only intended to accept a string or Text component;
-   * other use cases, while allowed, are not supported and may result in unexpected behavior.
+   * other use cases, while allowed, are not supported and may result in unexpected behavior. For arbitrary content, use `subdetailNode`.
    */
   subdetail?: React.ReactNode;
+  /** React node to render subdetail. Takes precedence over `subdetail`. */
+  subdetailNode?: React.ReactNode;
   /** Variant color to apply to the subdetail text. */
   variant?: CellDetailVariant;
   /** Specifies whether font should be scaled down automatically to fit given style constraints. */
@@ -34,19 +38,25 @@ const variantColorMap: Record<CellDetailVariant, ThemeVars.Color> = {
 export const CellDetail = memo(function CellDetail({
   adjustsFontSizeToFit,
   detail,
+  detailNode,
   subdetail,
+  subdetailNode,
   variant = 'foregroundMuted',
   subdetailFont = 'label2',
 }: CellDetailProps) {
   return (
     <VStack alignContent="flex-end" alignItems="flex-end" justifyContent="center">
-      {!!detail && (
+      {detailNode ? (
+        detailNode
+      ) : detail ? (
         <Text adjustsFontSizeToFit={adjustsFontSizeToFit} font="body" numberOfLines={1}>
           {detail}
         </Text>
-      )}
+      ) : null}
 
-      {!!subdetail && (
+      {subdetailNode ? (
+        subdetailNode
+      ) : subdetail ? (
         <Text
           adjustsFontSizeToFit={adjustsFontSizeToFit}
           color={variantColorMap[variant]}
@@ -55,7 +65,7 @@ export const CellDetail = memo(function CellDetail({
         >
           {subdetail}
         </Text>
-      )}
+      ) : null}
     </VStack>
   );
 });