+++ packages/web/src/cells/CellDetail.tsx
@@ -10,14 +10,18 @@ export type CellDetailVariant = 'foregroundMuted' | 'negative' | 'positive' | 'w
 export type CellDetailProps = {
   /**
    * Label and/or extra detail. This prop is only intended to accept a string or Text component;
-   * other use cases, while allowed, are not supported and may result in unexpected behavior.
+   * other use cases, while allowed, are not supported and may result in unexpected behavior. For arbitrary content, use `detailNode`.
    */
   detail?: React.ReactNode;
+  /** React node to render label and/or extra detail. Takes precedence over `detail`. */
+  detailNode?: React.ReactNode;
   /**
    * Subdetail providing more information. This prop is only intended to accept a string or Text component;
-   * other use cases, while allowed, are not supported and may result in unexpected behavior.
+   * other use cases, while allowed, are not supported and may result in unexpected behavior. For arbitrary content, use `subdetailNode`.
    */
   subdetail?: React.ReactNode;
+  /** React node to render subdetail. Takes precedence over `subdetail`. */
+  subdetailNode?: React.ReactNode;
   /** Variant color to apply to the subdetail text. */
   variant?: CellDetailVariant;
   /** Font to apply to the subdetail text. */
@@ -36,16 +40,22 @@ export const CellDetail = memo(function CellDetail({
   subdetail,
   variant = 'foregroundMuted',
   subdetailFont = 'label2',
+  detailNode,
+  subdetailNode,
 }: CellDetailProps) {
   return (
     <>
-      {!!detail && (
+      {detailNode ? (
+        detailNode
+      ) : detail ? (
         <Text as="div" display="block" font="body" overflow="truncate" textAlign="end">
           {detail}
         </Text>
-      )}
+      ) : null}
 
-      {!!subdetail && (
+      {subdetailNode ? (
+        subdetailNode
+      ) : subdetail ? (
         <Text
           as="div"
           color={variantColorMap[variant]}
@@ -56,7 +66,7 @@ export const CellDetail = memo(function CellDetail({
         >
           {subdetail}
         </Text>
-      )}
+      ) : null}
     </>
   );
 });